agriscan-ai/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.1
pydantic==2.5.2
pydantic-settings==2.1.0
requests==2.31.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1
websockets==12.0
pillow==10.1.0
numpy==1.26.2
python-dotenv==1.0.0
